qmemman: round prefmem
There is no sense operating on fractional bytes here.
In practice, it should be rounded to whole pages or even more, but that
require few more changes.

+++ qubes/qmemman/algo.py
@@ -54,10 +54,10 @@ def prefmem(domain):
     # dom0 is special, as it must have large cache, for vbds. Thus, give it
     # a special boost
     if domain.id == '0':
-        return min(domain.mem_used * CACHE_FACTOR + DOM0_MEM_BOOST,
-            domain.memory_maximum)
-    return max(min(domain.mem_used * CACHE_FACTOR, domain.memory_maximum),
-        MIN_PREFMEM)
+        return int(min(domain.mem_used * CACHE_FACTOR + DOM0_MEM_BOOST,
+            domain.memory_maximum))
+    return int(max(min(domain.mem_used * CACHE_FACTOR, domain.memory_maximum),
+        MIN_PREFMEM))
 
 
 def memory_needed(domain):
